# api/openapi.yaml
openapi: 3.0.3
info:
  title: Switchyard Local API
  version: 1.0.0
  description: Local control-plane contract shared by browser, desktop, and CLI clients.
servers:
  - url: /api/v1
paths:
  /system:
    get:
      operationId: getSystem
      summary: Read daemon and storage status
      responses:
        "200":
          description: Current daemon status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemInfo"
        default:
          $ref: "#/components/responses/Problem"
  /host:
    get:
      operationId: getHost
      summary: Read current host CPU, memory, and aggregate Docker storage
      responses:
        "200":
          description: Current explicitly partial host observation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostObservation"
        default:
          $ref: "#/components/responses/Problem"
...
  /operations/{operationId}:
    parameters:
      - $ref: "#/components/parameters/OperationId"
    get:
      operationId: getOperation
      summary: Read a durable operation
      responses:
        "200":
          description: Durable operation state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        default:
          $ref: "#/components/responses/Problem"
  /operations:
    get:
      operationId: listOperations
      summary: List recent durable operations
      parameters:
        - name: projectId
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, format: int64, minimum: 1, maximum: 500, default: 100 }
      responses:
        "200":
          description: Recent durable operations
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Operation" }
        default:
          $ref: "#/components/responses/Problem"
  /operations/{operationId}/cancel:
    parameters:
      - $ref: "#/components/parameters/OperationId"
    post:
      operationId: cancelOperation
      summary: Request idempotent operation cancellation
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Current operation after the cancellation request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        default:
          $ref: "#/components/responses/Problem"
...
  /projects:
    get:
      operationId: listProjects
      summary: List registered projects
      responses:
        "200":
          description: Registered project catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        default:
          $ref: "#/components/responses/Problem"
  /projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      operationId: getProject
      summary: Read one registered project
      responses:
        "200":
          description: Registered project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        default:
          $ref: "#/components/responses/Problem"
    delete:
      operationId: removeProject
      summary: Remove catalog state without changing repository files
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "204": { description: Project removed }
        default:
          $ref: "#/components/responses/Problem"
  /projects/{projectId}/trust:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      operationId: trustProject
      summary: Validate and accept the latest project proposal
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Trusted project and accepted proposal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AcceptedManifestProposal" }
        default:
          $ref: "#/components/responses/Problem"
...
  /projects/{projectId}/environments:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      operationId: listProjectEnvironments
      summary: List registered Git worktree environments for a trusted project
      responses:
        "200":
          description: Stable worktree environment registrations
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProjectEnvironment" }
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: registerProjectEnvironments
      summary: Reconcile trusted Git worktrees and allocate exact runtime ports
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Complete worktree registration result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvironmentRegistration" }
        default:
          $ref: "#/components/responses/Problem"
  /environments/{environmentId}:
    parameters:
      - name: environmentId
        in: path
        required: true
        schema: { type: string, minLength: 1, maxLength: 128 }
    get:
      operationId: getEnvironment
      summary: Read one registered project environment
      responses:
        "200":
          description: Registered environment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectEnvironment" }
        default:
          $ref: "#/components/responses/Problem"
    patch:
      operationId: updateEnvironment
      summary: Update the friendly localhost hostname without changing runtime state
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EnvironmentUpdate" }
      responses:
        "200":
          description: Updated environment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectEnvironment" }
        default:
          $ref: "#/components/responses/Problem"
  /routes:
    get:
      operationId: listLocalRoutes
      summary: List optional localhost routes and explicit availability states
      responses:
        "200":
          description: Current local route registry
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LocalRoute" }
        default:
          $ref: "#/components/responses/Problem"
...
components:
  parameters:
    OperationId:
      name: operationId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  responses:
    Problem:
      description: RFC 9457-style problem details
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
  schemas:
...
    Project:
      type: object
      additionalProperties: false
      required: [id, slug, displayName, trustState, primaryLocation, tags, manifestRevision, createdAt, updatedAt]
      properties:
        id: { type: string }
        slug: { type: string }
        displayName: { type: string }
        description: { type: string }
        trustState:
          type: string
          enum: [pending, trusted, rejected]
        primaryLocation: { type: string }
        tags:
          type: array
          items: { type: string }
        manifestRevision: { type: integer, format: int64, minimum: 0 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    SourceRange:
      type: object
      additionalProperties: false
      required: [startLine, endLine]
      properties:
        startLine: { type: integer, minimum: 1 }
        endLine: { type: integer, minimum: 1 }
    DiscoveryEvidence:
      type: object
      additionalProperties: false
      required: [id, scanner, kind, sourcePath, location, confidence, data, warnings]
      properties:
        id: { type: string }
        scanner: { type: string }
        kind: { type: string }
        sourcePath: { type: string }
        location: { $ref: "#/components/schemas/SourceRange" }
        confidence: { type: number, format: double, minimum: 0, maximum: 1 }
        data:
          type: object
          additionalProperties: true
        warnings:
          type: array
          items: { type: string }
    ManifestValidation:
      type: object
      additionalProperties: false
      required: [valid, errors, warnings]
      properties:
        valid: { type: boolean }
        errors:
          type: array
          items: { type: string }
        warnings:
          type: array
          items: { type: string }
    ManifestProposal:
      type: object
      additionalProperties: false
      required: [id, projectId, scannerVersion, schemaVersion, candidate, evidence, confidenceByField, unresolved, validation, status, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        scannerVersion: { type: string }
        schemaVersion: { type: string }
        candidate:
          type: object
          description: Canonical document governed by internal/manifest/schema/project.schema.json.
          additionalProperties: true
        evidence:
          type: array
          items: { $ref: "#/components/schemas/DiscoveryEvidence" }
        confidenceByField:
          type: object
          additionalProperties: { type: number, format: double }
        unresolved:
          type: array
          items: { type: string }
        validation: { $ref: "#/components/schemas/ManifestValidation" }
        status:
          type: string
          enum: [proposed, accepted, rejected, superseded]
        createdAt: { type: string, format: date-time }
    AcceptedManifestProposal:
      type: object
      additionalProperties: false
      required: [project, proposal]
      properties:
        project: { $ref: "#/components/schemas/Project" }
        proposal: { $ref: "#/components/schemas/ManifestProposal" }
...
    EnvironmentPortLease:
      type: object
      additionalProperties: false
      required: [portId, protocol, targetPort, hostPort]
      properties:
        portId: { type: string }
        protocol: { type: string, enum: [tcp, udp] }
        targetPort: { type: integer, minimum: 1, maximum: 65535 }
        hostPort: { type: integer, minimum: 1, maximum: 65535 }
    EnvironmentRuntimeAllocation:
      type: object
      additionalProperties: false
      required: [composeProjectName, portLeaseNamespace, portOffset, portLeases]
      properties:
        composeProjectName: { type: string }
        portLeaseNamespace: { type: string }
        portOffset: { type: integer, minimum: 1, maximum: 65535 }
        portLeases:
          type: array
          items: { $ref: "#/components/schemas/EnvironmentPortLease" }
    ProjectEnvironment:
      type: object
      additionalProperties: false
      required: [id, projectId, name, path, detached, bare, locked, primary, availability, state, hostname, allocation, registeredAt, lastObservedAt, updatedAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        name: { type: string }
        path: { type: string }
        head: { type: string }
        branch: { type: string }
        detached: { type: boolean }
        bare: { type: boolean }
        locked: { type: boolean }
        primary: { type: boolean }
        availability: { type: string, enum: [available, unavailable] }
        unavailableReason: { type: string }
        state: { type: string, enum: [registered, active, inactive, unavailable] }
        hostname: { type: string }
        target: { type: string }
        allocation: { $ref: "#/components/schemas/EnvironmentRuntimeAllocation" }
        registeredAt: { type: string, format: date-time }
        lastObservedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    EnvironmentRegistration:
      type: object
      additionalProperties: false
      required: [projectId, environments, removedIds, observedAt]
      properties:
        projectId: { type: string }
        environments:
          type: array
          items: { $ref: "#/components/schemas/ProjectEnvironment" }
        removedIds:
          type: array
          items: { type: string }
        observedAt: { type: string, format: date-time }
    EnvironmentUpdate:
      type: object
      additionalProperties: false
      required: [hostname]
      properties:
        hostname: { type: string, minLength: 1, maxLength: 253 }
    LocalRoute:
      type: object
      additionalProperties: false
      required: [hostname, status, candidateEnvironmentIds, updatedAt]
      properties:
        hostname: { type: string }
        status: { type: string, enum: [active, unavailable, conflict, disabled] }
        projectId: { type: string }
        environmentId: { type: string }
        target: { type: string }
        reason: { type: string }
        candidateEnvironmentIds:
          type: array
          items: { type: string }
        updatedAt: { type: string, format: date-time }
...
    Operation:
      type: object
      additionalProperties: false
      required:
        - id
        - projectId
        - kind
        - state
        - requestedAt
        - updatedAt
        - cancellationRequested
      properties:
        id:
          type: string
        projectId:
          type: string
        workspaceId:
          type: string
        kind:
          type: string
        state:
          $ref: "#/components/schemas/OperationState"
        errorCode:
          type: string
        errorMessage:
          type: string
        cancellationRequested:
          type: boolean
        requestedAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OperationState:
      type: string
      enum: [queued, running, succeeded, failed, cancelled, partially_succeeded]
    SystemInfo:
      type: object
      additionalProperties: false
      required:
        - status
        - version
        - commit
        - apiVersion
        - databaseSchemaVersion
        - startedAt
      properties:
        status:
          type: string
          enum: [ready]
        version:
          type: string
        commit:
          type: string
        builtAt:
          type: string
          format: date-time
          nullable: true
        apiVersion:
          type: string
        databaseSchemaVersion:
          type: integer
          format: int64
          minimum: 0
        startedAt:
          type: string
          format: date-time
    DockerHostObservation:
      type: object
      additionalProperties: false
      required: [connected, attribution]
      properties:
        connected: { type: boolean }
        storageBytes: { type: integer, format: int64, minimum: 0 }
        reclaimableBytes: { type: integer, format: int64, minimum: 0 }
        attribution: { type: string, enum: [shared, unknown] }
    HostObservation:
      type: object
      additionalProperties: false
      required: [cpuPercent, memoryUsedBytes, memoryTotalBytes, docker, observedAt, warnings]
      properties:
        cpuPercent: { type: number, format: double, minimum: 0 }
        memoryUsedBytes: { type: integer, format: int64, minimum: 0 }
        memoryTotalBytes: { type: integer, format: int64, minimum: 0 }
        docker: { $ref: "#/components/schemas/DockerHostObservation" }
        observedAt: { type: string, format: date-time }
        warnings:
          type: array
          items: { type: string }
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, code, correlationId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        code:
          type: string
        correlationId:
          type: string